src/psearch/ingestion_spanner/requirements.txt
# Google Cloud dependencies
google-cloud-bigquery>=3.11.4
google-cloud-spanner>=3.35.0
google-cloud-storage>=2.13.0
google-cloud-aiplatform>=1.36.4
google-genai>=0.1.0

# ML dependencies for vectorization
scikit-learn>=1.3.0
numpy>=1.24.0

# Vertex AI for Gemini
vertexai>=0.0.1

# Logging and utilities
python-dotenv>=1.0.0

# Added for requests
requests>=2.31.0

# Added for retry logic
tenacity>=8.2.0
